Fix subtraction operator type ambiguity in opacity calculation

+++ OurApp/OurApp/ImageUtilities.swift
@@ -191,7 +191,7 @@ struct FullScreenPhotoViewer: View {
     var body: some View {
         ZStack {
             Color.black
-                .opacity(1 - abs(dragOffset) / CGFloat(500))
+                .opacity(CGFloat(1) - abs(dragOffset) / CGFloat(500))
                 .ignoresSafeArea()
 
             TabView(selection: $currentIndex) {
